Refactor: Extract `unix/macos.rs`
